Add new pyenv segment to show active python version
This commit provides a new segment `pyenv` to show current python environment.
This prompt can be used in place of `virtualenv` segment. However, it is
targeted at users of [pyenv](https://github.com/yyuu/pyenv) and not meant to
replace the `virtualenv` segment.

The prompt works by parsing output of `pyenv version` and displaying the first
word of the output as segment text. The design (color etc.) is same as the
`virtualenv` segment including the icon. The segment would need to be
revisited if `pyenv version` changes its output format.

Tested on my personal machines (Mac OSX 10.11.4 and Ubuntu 15.04).