Hide one-space indent at right margin
ZSH displays the right prompt indented one space from the right
margin. This can be tweaked with ZLE_RPROMPT_INDENT=0, but that
makes display problems very likely and I don't recommend it. But
all is not lost.

The %E prompt format does continue the current background color
through that last indent space. So it can *look* like our prompt
goes right up to the margin, if:
1) We remove the trailing space from the last segment, and
2) We add %E before clearing the background color