set DISABLE_UPDATE_PROMPT=true when activating instant prompt
This is an Oh My Zsh option that makes it update without asking for
confirmation. In practice, when instant prompt is active, OMZ doesn't
wait for the user confirmation and automatically updates.

By setting DISABLE_UPDATE_PROMPT=true we get rid of the spurious
question from the console output without changing the behavior of OMZ
updater.