GCloud Segment shows proj name not proj ID.
Logic behind change: gcloud project IDs are semi randomly generated and can not be changed by the user. The project name can be changed and is much easier on the eyes, and easier for the user the identify the project, as it will be what they typed in when naming the project, as opposed to the ID which they won't have typed in.