Fixes "bad floating point" problem with task-3.0.1
After upgrading TaskWarrior to 3.0.1, I think... the comand means to
return some "Runnable" which resolves to a float, now returns the float
itself. So when we try to resolve the Runnable, ZSH throws an exception,
and p10k cannot parse the float from that. This just changes the
assumption that we get a float back.

p.s: Please when reviewing, ask youself "How does this affect P10K at
large?" I don't have this purview, but this change sure fixes my
problem.

Fixes #2648